+++ src/devteam/__init__.py
@@ -1,5 +1,4 @@
-from .crew import VirtualCrew
-from .managers import ProjectManager
+from .crew import VirtualCrew, ProjectManager
 from .utils import LLMFactory
 
 __all__ = [